+++ low-level/files/opt/seagate/sspl/setup/sspl_cleanup.py
@@ -1,0 +1,109 @@
+#!/bin/env python3
+
+# Copyright (c) 2020 Seagate Technology LLC and/or its Affiliates
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
+# For any questions about this software or licensing,
+# please email opensource@seagate.com or cortx-questions@seagate.com.
+
+import shutil
+import os
+import pwd
+
+from cortx.utils.conf_store import Conf
+from cortx.utils.process import SimpleProcess
+from framework.utils.file_utils import FileUtils
+from framework.utils.service_logging import logger
+from framework.utils.utility import Utility
+from framework.base.sspl_constants import (
+    SSPL_CONFIG_INDEX, file_store_config_path, global_config_file_path,
+    sspl_config_path, PRODUCT_BASE_DIR, SSPL_BASE_DIR, PRODUCT_FAMILY,
+    RSYSLOG_IEM_CONF, RSYSLOG_SSPL_CONF, RSYSLOG_MSB_CONF,
+    IEM_LOGROTATE_CONF, SSPL_LOGROTATE_CONF, MSB_LOGROTATE_CONF,
+    RSYSLOG_SB_CONF, SB_LOGROTATE_CONF, USER)
+
+
+class SSPLCleanup:
+    """Reset config and optionally factory operation."""
+
+    def __init__(self, args):
+        self.pre_factory = False
+        if "--pre-factory" in args:
+            self.pre_factory = True
+
+    def validate(self):
+        pass
+
+    def process(self, product):
+        """Reset and cleanup config."""
+        try:
+            if os.path.exists(file_store_config_path):
+                FileUtils.reset_log_files(file_store_config_path, del_file=True)
+            shutil.copyfile("%s/conf/sspl.conf.%s.yaml" % (
+                SSPL_BASE_DIR, product), file_store_config_path)
+            if self.pre_factory:
+                self.cleanup_log_and_config()
+        except OSError as e:
+            logger.error(f"Failed in Cleanup. ERROR: {e}")
+
+    def cleanup_log_and_config(self):
+        """--pre-factory cleanup : Cleanup logs, config files and
+        undo everything whatever was done in post-install Mini-Provisioner
+        Interface."""
+
+        Conf.load(SSPL_CONFIG_INDEX, sspl_config_path)
+        sspl_log_file_path = Utility.get_config_value(
+            SSPL_CONFIG_INDEX, "SYSTEM_INFORMATION>sspl_log_file_path")
+        iem_log_file_path = Utility.get_config_value(
+            SSPL_CONFIG_INDEX, "IEMSENSOR>log_file_path")
+
+        # Directories and file which needs to deleted.
+        directories = [
+            f'/var/{PRODUCT_FAMILY}/sspl', f'/var/{PRODUCT_FAMILY}/iem/',
+            f'/var/log/{PRODUCT_FAMILY}/sspl/', f'/var/log/{PRODUCT_FAMILY}/iem/',
+            '/etc/sspl-ll/', f'{PRODUCT_BASE_DIR}/iem/iec_mapping']
+
+        sspl_sudoers_file = '/etc/sudoers.d/sspl'
+        sspl_dbus_policy_rules = '/etc/polkit-1/rules.d'
+        sspl_dbus_policy_conf = '/etc/dbus-1/system.d'
+        sspl_service_file = '/etc/systemd/system/sspl-ll.service'
+        sspl_test_backup = '/etc/sspl_tests.conf.back'
+        sspl_test_file_path = '/etc/sspl_test_gc_url.yaml'
+        sspl_sb_log_file_path = sspl_log_file_path.replace(
+            "/sspl.log", "/sspl_support_bundle.log")
+        manifest_log_file_path = sspl_log_file_path.replace(
+            "/sspl.log", "/manifest.log")
+        # symlinks created during post_install
+        sspl_ll_cli = "/usr/bin/sspl_ll_cli"
+        # Remove SSPL config other config/log files which we have
+        # created during post_install.
+        for filepath in [
+            sspl_ll_cli, sspl_test_backup, sspl_test_file_path,
+            file_store_config_path, global_config_file_path, sspl_log_file_path,
+            iem_log_file_path, sspl_sb_log_file_path, manifest_log_file_path,
+            RSYSLOG_IEM_CONF, RSYSLOG_SSPL_CONF, RSYSLOG_MSB_CONF,
+            IEM_LOGROTATE_CONF, SSPL_LOGROTATE_CONF, MSB_LOGROTATE_CONF,
+            RSYSLOG_SB_CONF, SB_LOGROTATE_CONF, sspl_dbus_policy_conf,
+                sspl_dbus_policy_rules, sspl_sudoers_file, sspl_service_file]:
+            FileUtils.reset_log_files(filepath, del_file=True)
+        # Delete directories which we have created during post_install.
+        for directory in directories:
+            FileUtils.reset_log_files(directory, del_dir=True)
+        logger.info("Deleted config/log files and directories.")
+        # Delete sspl-ll user
+        usernames = [x[0] for x in pwd.getpwall()]
+        if USER in usernames:
+            _, err, rc = SimpleProcess("/usr/sbin/userdel -f %s" % USER).run()
+            if rc != 0:
+                logger.info("Error occurref while deleteing %s user. ERROR: %s"
+                    %(USER, err))
+            else:
+                logger.info("Deleted %s user." % USER)
