+++ low-level/framework/utils/utility.py
@@ -28,7 +28,6 @@ import shutil
 import time
 from cortx.utils.conf_store import Conf
 from cortx.utils.conf_store.error import ConfError
-from cortx.utils.process import SimpleProcess
 from framework.utils.service_logging import logger
 
 
@@ -167,33 +166,6 @@ class Utility(object):
     def get_current_time():
         """Returns the time as integer number in seconds since the epoch in UTC."""
         return int(time.time())
-
-    @staticmethod
-    def reset_log_files(path, fformat=None, del_file=False, del_dir=False):
-        """Clean log files and delete files from dir."""
-        if not os.path.exists(path):
-            logger.info(f"{path} path doesn't exists.")
-            return
-        if fformat:
-            for root, _, files in os.walk(path):
-                for file in files:
-                    if file.endswith(fformat):
-                        if del_file:
-                            os.remove(os.path.join(root, file))
-                            return
-                        cmd = f"truncate -s 0 > {os.path.join(root, file)}"
-                        _, error, returncode = SimpleProcess(cmd).run()
-                        if returncode != 0:
-                            logger.error(
-                                "Failed to clear log file data. "
-                                f"ERROR:{error} CMD:{cmd}")
-        else:
-            if os.path.isfile(path) and del_file:
-                os.remove(path)
-            elif os.path.islink(path):
-                os.unlink(path)
-            elif os.path.isdir(path) and del_dir:
-                shutil.rmtree(path, ignore_errors=True)
 
 
 def errno_to_str_mapping(err_no):
